UserController + Basic AuthController
Created the UserController. Users can retrieve, update, and delete data about themselves. Following to the creation of the UserController, there will be an AdminUserController where an admin can interact with user data.
Also implemented a basic AuthController where a user can be created and a user can login. Following to the creation of AuthController bearer auth will be implemented to restrict endpoint access when needed.